CHROMIUM: arm: tegra: Round up PWM when appropriate
Sometimes the PWM duty cycle ends up being one less than it should be,
due to rounding errors.  Using our current case as an example:

  period=4750000
  max=255
  pwm_backlight_update_status() is called with brightness=10.
  brightness is scaled to floor(10 * 4750000 / 255) = 186274.

  Then pwm_config() is called with duty_ns=186274.
  The calculated PWM value is:
     floor(duty_ns * 255 / 4750000)
   = floor(9.9999726316)
   = 9

The value should be rounded up to 10.  This patch modifies the pwm code
to round to the nearest integer, rather than automatically round down.

BUG=chrome-os-partner:2603
TEST=In a dark room, type the following:
  > stop powerd
  > echo 255 > /sys/class/backlight/pwm-backlight/brightness
  > echo 1 > /sys/class/backlight/pwm-backlight/brightness
The backlight should be very dim but still on, and the text on screen
should be visible in the dark room.

Change-Id: Ibc9c2b0cfd937209561a8b46727e703ee4dfedf4
Signed-off-by: Simon Que <sque@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/8715
Reviewed-by: Olof Johansson <olofj@chromium.org>
Reviewed-by: Bryan Freed <bfreed@chromium.org>