CHROMIUM: efi: Only create efi_smbios_base if it exists
Our dmidecode gets the smbios base address from
/sys/kernel/debug/efi_smbios_base. If that file exists, dmidecode
will not fall back at using the default smbios locations. Unfortunately
the file is even created on non-EFI systems, where the EFI data
structure is not filled with valid information. This patch changes the
code so it only creates the file if there is a non-0 value for
efi_smbios_base (because the smbios base can not live at address 0
anyways)

Signed-off-by: Stefan Reinauer <reinauer@google.com>

BUG=chrome-os-partner:4197
TEST=boot with coreboot, see dmidecode starts working

Change-Id: I60f7ac5caa0b01c34b0891db2b4bf5c4adbd3e18
Reviewed-on: http://gerrit.chromium.org/gerrit/8716
Tested-by: Stefan Reinauer <reinauer@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>