CHROMIUM: USB: Enable wake for host controllers
This patch is a result of upstream discussion regarding enabling USB
wake by default for intermediate devices.  The patch came from Alan
Stern here:
http://marc.info/?l=linux-usb&m=131614057108460&w=2

Further discussions on changing default wakeup policy triggered by the
above discussion are happening here:
http://marc.info/?t=131613872700001&r=1&w=2

There seems to consensus upstream on the change to default wakeup policy
although the patch itself has not made it through the upstream trees
yet. Not marking this patch as UPSTREAM yet due to this.

BUG=chrome-os-partner:4459
TEST=Transition system to suspend-to-RAM state. On systems where power
is maintained to USB rails in this state, a USB keyboard should be able
to wake the system up.  Note that mouse devices will not be able to wake
the system just yet since wakeup is enabled by default only for
keyboard devices in the kernel.

Change-Id: I335c882683d5192f721ee6a610248ae9503bc227
Signed-off-by: Sameer Nanda <snanda@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/8912
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>