msm: sdio: Keep certain messages in a local buffer
While logging to the kernel log, keep the last 20 or so messages in
a local buffer. This can be accessed by debugfs or through a crash
dump. This is useful when there is heavy logging to the kernel log
and sdio messages are lost.

Signed-off-by: Krishna Konda <kkonda@codeaurora.org>
(cherry picked from commit c6e7d74e78b1eac960b9cebee46c7ff8d141004b)

Conflicts:

	arch/arm/mach-msm/sdio_al.c

Change-Id: Id669161ae533e88e3b876d2145377c68be316332
Signed-off-by: Shruthi Krishna <skrish@codeaurora.org>