drm/i915: export a CPT mode set verification function
At the point where we check, we can't do much about the failure, but it
can aid debugging.  Note that the auto-train override bit will be reset
as part of normal mode setting with this patch if a pipe ever does get
stuck, but that's consistent with the workaround for CPT provided by the
hardware team.  This patch helped catch the fact that the pipe wasn't
running in the !composite sync FDI case on my IVB SDV, so has already
shown to be useful.

Signed-off-by: Jesse Barnes <jbarnes@virtuousgeek.org>
Tested-By: Eugeni Dodonov <eugeni.dodonov@intel.com>
Reviewed-By: Eugeni Dodonov <eugeni.dodonov@intel.com>
Signed-off-by: Keith Packard <keithp@keithp.com>
(cherry picked from commit d4270e57efe9e2536798c59e1ed2fd0a1e5cdfcf)

Signed-off-by: Stéphane Marchesin <marcheu@chromium.org>

Conflicts:

	drivers/gpu/drm/i915/intel_drv.h

Change-Id: I0c962ab09528eadc0d84981e73d4fa803a7c7135
Reviewed-on: https://gerrit.chromium.org/gerrit/13301
Tested-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>