msm: rpc: While powering down flush all the rpcrouter workqueues
When the device is powered down, RPC Router closes all the transports
and releases all the associated data structures. This sometimes lead to
works queued in the RPC router workqueues to access stale data and hence
cause kernel panic. So when the device is powered down, abort the works
queued in the workqueues and wait until the workqueues are flushed and
then free the associated data structures.

CRs-Fixed: 312078
Signed-off-by: Karthikeyan Ramasubramanian <kramasub@codeaurora.org>
(cherry picked from commit c74ab97883834479ab1b39586469a36357e4323b)

Change-Id: If8511045e08b6c54a463df72900452e4c05f7ae6
Signed-off-by: Sunil Joseph <sunilj@codeaurora.org>