UPSTREAM: ALSA: hda - Cache the jack-detection value
Introduce a table containing the pins and their jack-detection states
for avoiding the unnecessary verbs to check the pin status at each time.

When the unsol event is enabled via snd_hda_jack_detect_enable(), it
automatically adds the given NID to the table.  Then the driver supposes
that the codec driver will set the dirty flag appropariately when an
unsolicited event is invoked for that pin.

The behavior for reading other pins that aren't registered in the table
doesn't change.  Only the pins assigned to the table are cached, so far.

In near futre, this table can be extended to use the central place for
the unsolicited events of all pins, etc, and eventually include the
jack-detect kcontrols that replace the current input-jack stuff.

BUG=chromium-os:20620
TEST=Manual, run kernel on stumpy and check for JACK controls.

Signed-off-by: Takashi Iwai <tiwai@suse.de>
(cherry picked from commit 1835a0f9a2121ce3198dab67507d4d3e960cc09e)

Change-Id: Ia90c8ec8699a4aa7462c58b7d98823136488975b
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/20715