CHROMIUM: ramoops: remove module parameters
The ramoops driver is intended to be used with platforms that define
persistent memory regions. If memory regions were configurable with
module parameters, it would be possible to read some RAM regions via
the new /proc interface without access to /dev/mem (which would
result in a loss of kernel memory privacy when a system is built with
STRICT_DEVMEM), so remove this ability completely.

BUG=chromium-os:12059
TEST=x86-alex build and boot, manually verified /dev/pstore/dmesg-ramoops-N
    showing breakme crash reports.

Change-Id: I5177af5b0a237b01633a9a202172504c8564c914
Signed-off-by: Kees Cook <keescook@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/11618