CHROMIUM: alsa: hda: cirrus: Pin config for Stumpy PVT.
For the Stumpy PVT build the mic input has been relocated to the line
in input.  Reconfigure the pin verbs to reflect this.  Also fix the
name of the stumpy enum value to match the convention.

NOTICE: this change will break recording from mic in on DVT Stumpy
systems that have not been reworked. The mic input will disappear.

BUG=chrome-os-partner:6135
TEST=Must be tested on a PVT or newer stumpy. I used arecord and aplay
to verify this on a DVT board modified to match PVT.

Signed-off-by: Dylan Reid <dgreid@chromium.org>
Change-Id: If10a4716e21d7977d8f7d4f6547d1e31c5496ac4
Reviewed-on: https://gerrit.chromium.org/gerrit/11192
Reviewed-by: Sameer Nanda <snanda@chromium.org>

[3.2 FIXUP : enum STUMPY_CDB4210 changed to CS421X_STUMPY -grundler]