Input: cyapa - refactor finger payload processing
Per Henrik Rydberg feedback:
  Since the set of tracking ids is so limited, the simplest implementation
is to equate the device tracking ids with the MT slots. A book-keeping bit
field can be used to keep track of what slots should be cleared in a round.

As a side affect, the whole cyapa_report_data abstraction goes away, as
well as the debug function cyapa_dump_report().

Also, per Alan Cox feedback:
  Pack on the wire data structs to ensure compatibility across arches.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chromium-os:21047
TEST=builds clean;
     motion, tap, and click should work as before with xf86-input-cmt

Change-Id: I25f456885b152d6e86ed3cadf02c044633753721
Reviewed-on: https://gerrit.chromium.org/gerrit/10968
Reviewed-by: Benson Leung <bleung@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>