drivers: i915: Fix BLC PWM register setup
There is an error in i915_read_blc_pwm_ctl, where the register values
are not being copied correctly.  BLC_PWM_CTL and BLC_PWM_CTL2 are
getting mixed up.  This patch fixes that so that saveBLC_PWM_CTL2 and
not saveBLC_PWM_CTL is copied to the BLC_PWM_CTL2 register.

BUG=chrome-os-partner:5570
TEST=backlight must work

Change-Id: I2bccac5fbc305d8b4c91c6478b3e07aa43cfc101
Signed-off-by: Simon Que <sque@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/11673
Reviewed-by: Sameer Nanda <snanda@chromium.org>
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>