CHROMIUM: Input: cyapa - protect bootloader entry/exit with mutex
Use a mutex to protect entry and exit from bootloader.

The CYAPA bootloader differs from operation mode in two key respects.
  1) it uses a different, incompatible register set that can only be
     accessed with i2c block read commands (even for smbus devices).
  2) it doesn't use interrupts

The implication of (1) is that the driver must take care to never try to
access op mode registers when the device is in bootloader mode.  Such an
attempt will either be rejected by the bootloader, or return unexpected
values if the bootloader happens to support a read of an identical
register address.  Note that the driver uses such an read to determine the
current run mode.  For an i2c device, reading register 0 is valid in both
modes, but returns different data which can be used to distinguish them.

The implication of (2) is that irqs may be kept enabled when in the
bootloader, but that the irq handler should not attempt to read touch data
if such an in-bootloader irq occurs.  The irq should just be ignored.
Note that there is at least one such 'spurious' interrupt that must
always be ignored; when the device enters bootloader, it asserts its irq
line and keeps it high until it re-enters operation mode.

As a result of (1) and (2), there is a race when entering the bootloader.
Once the decision to enter bootloader has been taken, the only
allowed i2c commands are the soft_reset itself and bootloader commands.
In particular, the irq handler should not issue any more i2c requests.

To solve this race, a mutex is used to synchronize access to the state
variable in_bootloader. A mutex is used instead of a spinlock because
accessing the i2c bus can sleep, and sleeping is not permitted while
holding a spinlock.

As a side-affect of this:
 (a) the old "detect_status" becomes redundant and is removed.
 (b) there is no more cyapa_bl_enable/disable_irq() nor related variables.
 (c) recursive mutex lock is not allowed, therefore we only hold lock when
     calling a bus access function (e.g. lock in cyapa_get_query_data
     before calling cyapa_read_block), not when calling a helper function
     that calls bus access functions (e.g. not in cyapa_reconfig before
     calling cyapa_get_query_data).

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chromium-os:22608
TEST=Turn on kernel debugging.
  Boot, suspend/resume and enter/exit bootloader all work.
  Monitor dmesg while moving finger on touchpad.
  There should be no messages like this:
    [  129.686105] BUG: sleeping function called from invalid context at kernel/rtmutex.c
    [  129.701731] in_atomic(): 0, irqs_disabled(): 128, pid: 47, name: irq/370-cyapa
    [  129.708965] [<c00521bc>] (unwind_backtrace+0x0/0xec) from [<c043caf8>] (dump_stack+0x20/0x24)
    [  129.717487] [<c043caf8>] (dump_stack+0x20/0x24) from [<c0071f00>] (__might_sleep+0xec/0x10c)
    [  129.725921] [<c0071f00>] (__might_sleep+0xec/0x10c) from [<c04427e0>] (rt_mutex_lock+0x28/0x48)
    [  129.734622] [<c04427e0>] (rt_mutex_lock+0x28/0x48) from [<c03009b0>] (tegra_i2c_xfer+0x50/0x2d0)
    [  129.743398] [<c03009b0>] (tegra_i2c_xfer+0x50/0x2d0) from [<c02fea48>] (i2c_transfer+0x9c/0xec)
    [  129.752087] [<c02fea48>] (i2c_transfer+0x9c/0xec) from [<c02fef18>] (i2c_smbus_xfer+0x3c4/0x518)
    [  129.760860] [<c02fef18>] (i2c_smbus_xfer+0x3c4/0x518) from [<c02ff150>] (i2c_smbus_read_i2c_block_data+0x5c/0xa0)
    [  129.771107] [<c02ff150>] (i2c_smbus_read_i2c_block_data+0x5c/0xa0) from [<c02f901c>] (cyapa_read_block+0xe4/0x108)
    [  129.781441] [<c02f901c>] (cyapa_read_block+0xe4/0x108) from [<c02f9324>] (cyapa_irq+0x4c/0x1bc)
    [  129.790131] [<c02f9324>] (cyapa_irq+0x4c/0x1bc) from [<c00c221c>] (irq_thread+0xe0/0x208)
    [  129.798304] [<c00c221c>] (irq_thread+0xe0/0x208) from [<c009c0c4>] (kthread+0x98/0xa0)
    [  129.806220] [<c009c0c4>] (kthread+0x98/0xa0) from [<c004c910>] (kernel_thread_exit+0x0/0x8)

Change-Id: I6b3964805badcf46ead899bcfcd0a6ed4e007109
Reviewed-on: https://gerrit.chromium.org/gerrit/11320
Reviewed-by: Benson Leung <bleung@chromium.org>
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>