CHROMIUM: mmc: tegra: add delay on non-present card status
Seems that the card present detect needs a while to settle on tegra.
It used to luck out that resume took long enough that we didn't hit this,
but recently we are.

So, the case where card detect reports no card, we add a delay and
re-read. This is not on performance critical path (since the stack will
error out when the first failed card detect read happens), so we should
be ok from a performance perspective.

BUG=chromium-os:18948
TEST=suspend and resume, make sure machine doesn't crater

Change-Id: I176cdad13b1ae9eac537b8f788b3e15468cac3bd
Signed-off-by: Olof Johansson <olofj@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/11509
Reviewed-by: Benson Leung <bleung@chromium.org>