CHROMIUM: Input: cyapa - refactor detect_work
Probe detect and resume detect do essentially the same thing.  So, merge
them into a single function.
In both cases, the device is first determined to be fully operational.
The first time a detection happens, ->input will be NULL, and the input
device will be created.  On subsequent detections, full power will be
applied to the device.

Also, use PREPARE_WORK for the resume case, since the ->detect_work was
already initialized in probe().

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chromium-os:21047
TEST=builds, device detected on boot and after suspend/resume.

Change-Id: If17b6f988e82bc8fd6aeac6dbf81d328ea64a2e9
Reviewed-on: https://gerrit.chromium.org/gerrit/11614
Reviewed-by: Benson Leung <bleung@chromium.org>
Tested-by: Benson Leung <bleung@chromium.org>
Reviewed-by: Dudley Du <dudl@cypress.com>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>