mfd: timpani-codec: Making AUXPGA gain register as uncacheable
AUXPGA LR gain register is treated as a cacheable with which the
register value is changed when device is opened twice. So making
the LR gain register as uncacheable which in turn apply the gain
value to the codec every time device is opened.

CRs-Fixed: 319366
Signed-off-by: Vinay Vaka <vvaka@codeaurora.org>
(cherry picked from commit df2d0214edc5a59b5a8ec3002e3d6e86ece51466)

Change-Id: I2e1017e006e7d8abd663b95fe7a3a58b1eb76d02
Signed-off-by: Padakandla Phani Kumar <pphani@codeaurora.org>