CHROMIUM: Input: cyapa - refactor get_state
The BL_KEY checks actually show whether the bootloader detects valid
firmware or not.  This check is actually orthogonal to a device's
current operational state.  For instance, a device with erased memory
will boot into BL_IDLE state, but will not have BL_KEY bytes in is
BL_HEAD.

As a side affect, we only probe for the 3 status registers instead of
all 16 bytes of the BL_HEAD.  For smbus devices, we can use the
equivalent CYAPA_CMD_BL_STATUS command for reading these equivalent 3
registers if the i2c reg_read fails when in operational mode.

Also, since OP_DATA_VALID indicates that the device has data to send to
the host, it need not be checked when probing for device state.  If probing
happens very early when the device is starting its firmware, the device
will indicate its mode as DEV_NORMAL, but not yet have data to send, so
DATA_VALID is still 0.

This patch also adds some dev_dbg statements to highlight state transitions
while debugging.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chromium-os:23454
TEST=In following scenarios, check dmesg after boot to ensure device is
  detected properly:
  (1) cold boot with i2c device
  (2) cold boot with smbus device
  (3) warm boot with smbus device (no device power cycle during boot)
  (4) cold boot with i2c device + no firmware
  (5) cold boot with smbus device + no firmware
  (6) suspend/resume

Change-Id: I6bc964cb33fc42fc5e9fa955f8f3bf2095869ee3
Reviewed-on: https://gerrit.chromium.org/gerrit/12124
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>