CHROMIUM: Input: cyapa - refactor bootloader entry
The cyapa soft reset command only works in op mode, and we only use it
to enter the bootloader.  So, rename the function, and make it only send
a reset command if device is in op mode.

Also, since the driver registers as an input device when the device
enters operational mode.  Similarly, the device should unregister when
entering bootloader.  As a side affect of this, use the cyapa_detect()
path to properly re-register when exiting bootloader from
CYAPA_CMD_IDLE_TO_APP entry point of misc interface.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chromium-os:23454
TEST=builds clean;
  Use cyapa_fw_udpate -b -o /tmp/cyapa.bin to test enter/exit BL.
  Check /var/log/Xorg.0.log after reading firmware to confirm:
    X driver should unload, then reload, in response to udev events

Change-Id: Ie84ce47cfbc2371b55fc2eaaa0fe33175146e4f8
Reviewed-on: https://gerrit.chromium.org/gerrit/12125
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>