CHROMIUM: Input: cyapa - always call enable_irq_wake in suspend
Move the device status check into cyapa_set_power_mode.
This allows enable_irq_wake() to be called even when the device is not
operational mode.
This balances the enable/disable pair in the unexpected case that a
device is stuck in bootloader because it has bad firmware, and the system
is suspended.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chromium-os:21047
TEST=(1) Regression test: powerd_suspend
  => power mode should be set to LIGHT_SLEEP during suspend
  => power mode should be set to FULL_ACTIVE on resume
  (2) erase device firmware, then powerd_suspend
  => no attempt to change power mode; no kernel errors in dmesg

Change-Id: I0e2b12868c224c582c8a7eecb1ea51db332c8ffe
Reviewed-on: https://gerrit.chromium.org/gerrit/12123
Reviewed-by: Benson Leung <bleung@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>