CHROMIUM: HACK: ARM: tegra: Put CPU invalidation code in 2nd proc startup
We are seeing problems bringing up the 2nd processor on T25
machines when updating the kernel and doing a warm reboot.  That
points to the icache of the 2nd processor somehow keeping
stale state across a reboot.

As a temporary fix until someone can dig up the details, I'm
cramming the old 2.6.38 CPU invalidation code here.  It seems
to work, but without root cause it's a bit of a shot in the
dark.  NOTE: If there really are icache problems here, even
this isn't completely safe.  After all: who is to say that
there won't be stale icache data for tegra_secondary_startup()
and we'll never even make it to the icache flushing code.

BUG=chromium-os:23376
TEST=Ran reboot test described in above bug.

Change-Id: Icba16cffaa4e9efda326116f574fb9db67f1416b
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/12384
Reviewed-by: Olof Johansson <olofj@chromium.org>
Reviewed-by: Pawel Osciak <posciak@chromium.org>