CHROMIUM: Input: cyapa - only enable interrupts for device in op mode
Interrupts are only useful when a supported device is in operational mode
and the driver has already registered its input device.  However, it is
much easier to handle irq request errors in the driver probe() than in
detect_work().

So, we request_irq in probe(), but immediately disable it, and only
re-enable after registering the input device.

One side-affect of this change is that we disable_irq before unregistering
the input device when entering bootloader, and re-enable in the following
detect.  This actually solves the race condition between entering
bootloader and the irq handler, since disable_irq() itself is synchronous
- it only returns when the irq is disabled AND any running instance of the
handler has finished.  This, in turn, allows us to remove state_mutex
which we were using to accomplish the same thing.

Instead, there is another little race condition:
Since there is no way to request an irq that is initially disabled, there
is a brief window during probe(), between irq_request() and irq_disable(),
when an interrupt could occur (e.g., if user is touching pad during boot).
To resolve this, we check if cyapa->input in cyapa_irq() before doing
anything, since cyapa->input is NULL during this window, but valid for
all other proper interrupts.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chrome-os-partner:7058
TEST=boot system (with cyapa trackpad) while touching pad.
 There should be no messages like this in dmesg:
  input: cyapa as /devices/*/*-0067/input/input3
  Unable to handle kernel NULL pointer dereference...
  Internal error: Oops: 5 [#1] SMP
  Process irq/370-cyapa (pid: 40, stack limit = <>)
  Stack: (<> to <>)
  [<>] (input_event+0x2c/0x80) from [<>] (cyapa_irq+0xc8/0x1bc)
  [<>] (cyapa_irq+0xc8/0x1bc) from [<>] (irq_thread+0xe0/0x208)
  [<>] (irq_thread+0xe0/0x208) from [<>] (kthread+0x98/0xa0)
  [<>] (kthread+0x98/0xa0) from [<>] (kernel_thread_exit+0x0/0x8)
  ---[ end trace <> ]---
  exiting task "irq/370-cyapa" (40) is an active IRQ thread (irq 370)

Change-Id: I85b7782f3e0ecb831e41c459391cd6582752d643
Reviewed-on: https://gerrit.chromium.org/gerrit/12434
Reviewed-by: Benson Leung <bleung@chromium.org>
Reviewed-by: Dudley Du <dudl@cypress.com>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>