msm: board-qrd7627a: Enable BT/FM.
1) GPIO 114 on QRD is BT SYS RESET GPIO, which was
owned by modem proc, as a result set value/direction from
apps proc was not effective.
2) After using the correct gpio_tlmm_config api, apps proc
is able to own the gpio.
3) Once the GPIO is owned by apps proc, the set value/direction
for this GPIO are effective and as desired.
4) Correct BT_SYS_RESET GPIO functionality enables BT/FM SOC
to come out of RESET and perform as required.

Signed-off-by: Rahul Kashyap <rkashyap@codeaurora.org>
(cherry picked from commit 5e59b892f8031901cf373abb227cb32bcb53417b)

Change-Id: Ieb7c07aabc5e695718e3d261f605cf940d4dfc2c
Signed-off-by: Srivalli Oguri <oguri@codeaurora.org>