USB: msm_otg: Set PHCD bit if its not set by the controller during suspend
PHCD bit is automatically set by controller when SUSP bit is set in PORTSC
in Host mode. SUSP bit of PORTSC is set only when a device attached on the
root hub is suspended.

So during host mode suspend, set the PHCD bit if its not set by the
controller automatically.

Change-Id: I37aaab766d8f8eb658ae6fbeaf04be338e507fed
Signed-off-by: Rajkumar Raghupathy <raghup@codeaurora.org>