CHROMIUM: memconsole: Add ACPI method to locate cbmem start
With >=4GB memory config the memory map can be split and relying
on total memory for a physical address to scan near does not work.

Instead of relying on scanning memory there is now an NVS field
defined in coreboot to get the pointer to the first cbmem entry.

The existing memory walking method is left in place (for now?)
with one fix to check the pointer returned by ioremap.

BUG=chrome-os-partner:7350
TEST=Tested with Stumpy 2GB and 4GB config:

1) install latest coreboot firmware
2) install kernel with this patch
3) verify that /sys/firmware/log is valid

(cherry picked from commit e89d4132d87f2339233717cc3cb7174cad33e0da)

Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Change-Id: I05d7ab6011361a6a485aac9e03e3c604f07f9d47
Reviewed-on: https://gerrit.chromium.org/gerrit/13602
Reviewed-by: Stefan Reinauer <reinauer@chromium.org>