CHROMIUM: r8169: increase the delay parameter of pm_schedule_suspend
The r8169 driver in 3.1 kernel introduces power management to the driver.
Unfortunately the scheduled inactivity time is 100ms which causes the device to
oscillate between active and suspend when CAT5 is connected.  This patch
lengthens the inactivity delay to 5 seconds so that driver can establish
itself and remain active.

This patch is upstream in 3.2-rc4 but pure backport is non-trivial so
we'll need to carry the patch until then.

BUG=chrome-os-partner:6217
TEST=manual,

Run
     eth_pci_dev=`lspci | grep Ether | cut -d ' ' -f 1`
     while true ; do
       cat /sys/bus/pci/devices/0000\:${eth_pci_dev}/power/runtime_status
       sleep .5
     done

1. Make sure output is 'active' when CAT5 connected
2. Make sure output is 'suspended' when CAT5 disconnected and ~5secs have
elapsed.

Repeat for multiple connect/disconnect cycles

Signed-off-by: Todd Broch <tbroch@chromium.org>
Change-Id: I10e45a72be0631921004b54b825ea97ec5d57878
Reviewed-on: https://gerrit.chromium.org/gerrit/13500
Reviewed-by: Olof Johansson <olofj@chromium.org>