CHROMIUM: Input: cyapa - refactor set_power_mode
Early on, we would blindly attempt to set power mode to FULL power on
resume.  Since it was assumed the device may have been in bootloader mode,
there was a chance for the read_byte, as well as subsequent write_bytes
to fail.  So, read errors were ignored, and writing was done in a loop.

Now we only set power mode after a successful detect when we know we are
talking to a device in operational mode.  Thus, any read or write failures
are real bus errors that we should log like any other.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chromium-os:21047
TEST=powerd_suspend;
  device in deep sleep as measured by inductive monitoring of sample rate
  [tegra2_kaen]: device should wake system from S3
  device should return to full power on resume as measured by inductive
     monitoring of sample rate

Change-Id: Ie6e87a8073c2bd017462291ca6fd8eba496b59ed
Reviewed-on: https://gerrit.chromium.org/gerrit/13533
Reviewed-by: Dudley Du <dudl@cypress.com>
Reviewed-by: Benson Leung <bleung@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>